jest conversion and update for useAdministrationSubNav.test.tsx
